缓存重构及加载流程调整，加载失败尝试与db同步&类库相关依赖显示指定(4.5h)

+++ scirichon-cache/index.js
@@ -3,26 +3,23 @@ const rp = require('request-promise')
 const queryString = require('querystring')
 const config = require('config')
 const redis_config = config.get('redis')
-const RedisCache = require("node-cache-redis-fork");
+const RedisCache = require("node-cache-redis-fork")
+const common = require('scirichon-common')
+const schema = require('redis-json-schema')
+const uuid_validator = require('uuid-validate')
 
 const cache = new RedisCache({
     redisOptions: {host: redis_config.host, port: redis_config.port, db: 3},
     poolOptions: {priorityRange: 1}
 })
 
-const cmdb_type_routes = {
-    User: {route: '/users'},
-    ITService: {route: '/it_services/service'},
-    ITServiceGroup: {route: '/it_services/group'},
-    ServerRoom: {route: '/serverRooms'},
-    WareHouse: {route: '/wareHouses'},
-    Shelf: {route: '/shelves'},
-    Cabinet: {route: '/cabinets'},
-    ConfigurationItem: {route: '/cfgItems'},
-    ProcessFlow: {route: '/processFlows'}
-}
-
 const prefix = 'scirichon-cache:'
+
+let load_url = {}
+
+const setLoadUrl = (url)=>{
+    load_url = url
+}
 
 const set = async (key,val)=>{
     return await cache.set(prefix+key,val)
@@ -46,49 +43,53 @@ const flushAll = async ()=>{
     }
 }
 
-const internal_token_id = 'internal_api_invoke'
-
-const apiInvoker = function(method,url,path,params,body){
-    var options = {
-        method: method,
-        uri: url + path + (params?('?' + queryString.stringify(params)):''),
-        body:body,
-        json: true,
-        headers: {
-            'token': internal_token_id
+const saveItem = async (item)=>{
+    if(item&&item.uuid&&item.category){
+        if(item.category === 'User'){
+            await set(item.userid,{name:item.alias,uuid:item.userid,category:item.category,roles:item.roles})
+            await set(item.category + '_' + item.alias,{name:item.alias,uuid:item.userid,category:item.category,roles:item.roles})
+        }
+        else {
+            await set(item.uuid,{name:item.name,uuid:item.uuid,category:item.category,subtype:item.subtype})
+            if(item.name){
+                await set(item.category + '_' + item.name,{name:item.name,uuid:item.uuid,category:item.category,subtype:item.subtype})
+            }
         }
     }
-    return rp(options)
 }
 
-const loadAll = async (cmdb_url)=>{
-    await flushAll()
-    let results = [],key_id,key_name
-    for(let val of _.values(cmdb_type_routes)){
-        results.push(await apiInvoker('GET',cmdb_url,val.route,{'origional':true}))
-    }
-    for (let result of results){
-        result = result.data||result
-        if(result&&result.length){
-            for(let item of result){
-                if(item&&item.uuid){
-                    if(item.category === 'User'){
-                        key_id = item.userid,key_name = item.category + '_' + item.alias
-                        await set(key_id,{name:item.alias,uuid:item.userid,category:item.category})
-                        await set(key_name,{name:item.alias,uuid:item.userid,category:item.category})
-                    }
-                    else {
-                        key_id = item.uuid,
-                            await set(key_id,{name:item.name,uuid:item.uuid,category:item.category,subtype:item.subtype})
-                        if(item.name){
-                            key_name = item.category + '_' + item.name
-                            await set(key_name,{name:item.name,uuid:item.uuid,category:item.category,subtype:item.subtype})
-                        }
-                    }
+const loadAll = async ()=>{
+    let results = [],key_id,key_name,cmdb_type_routes = await schema.getApiRoutesAll()
+    if(load_url&&load_url.cmdb_url&&!_.isEmpty(cmdb_type_routes)){
+        await flushAll()
+        for(let val of _.values(cmdb_type_routes)){
+            results.push(await common.apiInvoker('GET',load_url.cmdb_url,val.route,{'origional':true}))
+        }
+        for (let result of results){
+            result = result.data||result
+            if(result&&result.length){
+                for(let item of result){
+                    await saveItem(item)
                 }
             }
         }
     }
+}
+
+const loadOne = async (category,uuid)=>{
+    let item,route,body,name,cmdb_type_routes = await schema.getApiRoutesAll()
+    if(load_url&&load_url.cmdb_url&&cmdb_type_routes[category]&&cmdb_type_routes[category].route) {
+        console.log(`load ${uuid} from cmdb`)
+        if(uuid_validator(uuid)||_.isInteger(uuid)){
+            body = {category,uuid,cypher:`MATCH (n:${category}) WHERE n.uuid={uuid} RETURN n`}
+        }else if(_.isString(uuid)){
+            name = uuid,body = {category,name,cypher:`MATCH (n:${category}) WHERE n.name={name} RETURN n`}
+        }
+        item = await common.apiInvoker('POST',load_url.cmdb_url,'/searchByCypher',{'origional':true},body)
+        item = item.data||item
+        await saveItem(item)
+    }
+    return item
 }
 
 const getByCategory = async (category)=>{
@@ -103,12 +104,20 @@ const getByCategory = async (category)=>{
 }
 
 const getItemByCategoryAndName = async (category,name)=>{
-    return await get(category+"_"+name)
+    let result = await get(category+"_"+name)
+    if(!result){
+        result = await loadOne(category,name)
+    }
+    return result
 }
 
 const getItemByCategoryAndID = async (category,uuid)=>{
-    return await get(uuid)
+    let result = await get(uuid)
+    if(!result){
+        result = await loadOne(category,uuid)
+    }
+    return result
 };
 
 
-module.exports = {loadAll,get,set,del,flushAll,getByCategory,getItemByCategoryAndName,getItemByCategoryAndID,cmdb_type_routes}
+module.exports = {loadAll,get,set,del,flushAll,getByCategory,getItemByCategoryAndName,getItemByCategoryAndID,setLoadUrl}
